Data loading screen: UI

Data loading screen: UI

data_loader_UI(id, app_title = "Multimorbidity Explorer")

Arguments

id
String with unique id of module in app

app_title
Name of your app. Defaults to "Multimorbidity Explorer"

Value

HTML component of shiny module

Examples

data_loader_UI('my_app')

#> <style type="text/css">/*
#> Fonts for app
#> */
#> 
#> body {
#>   --main-text-color: #525252;
#> }
#> 
#> body,
#> table,
#> label,
#> button,
#> svg text {
#>   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
#> }
#> 
#> 
#> h1, h2, h3, h4 {
#>   font-weight: 400;
#>   color: #252525;
#> }
#> 
#> p, span, table, label, button {
#>   color: var(--main-text-color);
#> }
#> 
#> svg text {
#>   fill: var(--main-text-color);
#> }
#> 
#> 
#> 
#> /* Button Styles */
#> button,
#> .btn {
#>   	-moz-box-shadow:inset 0px 1px 0px 0px #ffffff;
#>   	-webkit-box-shadow:inset 0px 1px 0px 0px #ffffff;
#>   	box-shadow:inset 0px 1px 0px 0px #ffffff;
#>   	background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #ffffff), color-stop(1, #f6f6f6));
#>   	background:-moz-linear-gradient(top, #ffffff 5%, #f6f6f6 100%);
#>   	background:-webkit-linear-gradient(top, #ffffff 5%, #f6f6f6 100%);
#>   	background:-o-linear-gradient(top, #ffffff 5%, #f6f6f6 100%);
#>   	background:-ms-linear-gradient(top, #ffffff 5%, #f6f6f6 100%);
#>   	background:linear-gradient(to bottom, #ffffff 5%, #f6f6f6 100%);
#>   	filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#ffffff', endColorstr='#f6f6f6',GradientType=0);
#>   	background-color:#ffffff;
#>   	-moz-border-radius:6px;
#>   	-webkit-border-radius:6px;
#>   	border-radius:6px;
#>   	border:1px solid #dcdcdc;
#>   	display:inline-block;
#>   	cursor:pointer;
#>   	padding:6px 20px;
#>   	text-decoration:none;
#>   	text-shadow:0px 1px 0px #ffffff;
#>   }
#> 
#> button:hover,
#> .btn:hover {
#> 	background:-webkit-gradient(linear, left top, left bottom, color-stop(0.05, #f6f6f6), color-stop(1, #ffffff));
#> 	background:-moz-linear-gradient(top, #f6f6f6 5%, #ffffff 100%);
#> 	background:-webkit-linear-gradient(top, #f6f6f6 5%, #ffffff 100%);
#> 	background:-o-linear-gradient(top, #f6f6f6 5%, #ffffff 100%);
#> 	background:-ms-linear-gradient(top, #f6f6f6 5%, #ffffff 100%);
#> 	background:linear-gradient(to bottom, #f6f6f6 5%, #ffffff 100%);
#> 	filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#f6f6f6', endColorstr='#ffffff',GradientType=0);
#> 	background-color:#f6f6f6;
#> }
#> 
#> button:active,
#> .btn:active {
#> 	position:relative;
#> 	top:1px;
#> }
#> 
#> span.btn {
#>   font-weight: 200;
#> }
#> 
#> button.btn {
#>   font-size: 1rem;
#>   font-weight: 400;
#> }</style>
#> <!DOCTYPE html>
#> <!-- template.html -->
#> 
#> <style>
#> 
#>   body {
#>     box-sizing: border-box;
#>     margin: 2px;
#>     background-color: #bdbdbd;
#>   }
#> 
#>   .info_header {
#>     height: 3rem;
#>   }
#> 
#>   .parent {
#>     height: calc(100vh - 3rem);
#>     display: grid;
#>     grid-template-columns: repeat(5, 1fr);
#>     grid-template-rows: repeat(5, 1fr);
#>     grid-column-gap: 15px;
#>     grid-row-gap: 15px;
#>   }
#> 
#>   .input-panel {
#>     grid-area: 1 / 1 / 6 / 3;
#>   }
#> 
#>   .instructions-panel {
#>     grid-area: 1 / 3 / 6 / 6;
#>   }
#> 
#>   .section {
#>     background-color: white;
#>     box-shadow: 3px 3px 5px 0px rgba(0,0,0,0.49);
#>     border-radius: 4px;
#>     overflow: hidden;
#>     padding: 1rem;
#>     overflow-y: scroll;
#>   }
#> 
#>   .app-title {
#>     margin: 0;
#>     padding-top: 0.5rem;
#>     padding-left: 0.5rem;
#>   }
#> 
#> 
#>   .form-group {
#>     padding: 0.5rem 0;
#>   }
#> 
#>   .form-group .input-group {
#>     padding-top: 0.5rem;
#>   }
#> 
#>   .form-group input {
#>     margin-left: 1rem;
#>     padding: 2px 5px;
#>   }
#> 
#>   .go-button {
#>     margin-top: 1.5rem;
#>   }
#> 
#> </style>
#> 
#> <div class = 'info_header'>
#>   <h2 class = 'app-title'> 
#> Multimorbidity Explorer
#>  </h2>
#> </div>
#> <div class = 'parent'>
#>   <div class = 'input-panel section'>
#>     
#> <div>
#>   <div id="my_app-preloaded_snps" class="shiny-html-output"></div>
#>   <h3>Load your data</h3>
#>   <div class="form-group shiny-input-container">
#>     <label class="control-label" for="my_app-phewas">Phewas results file</label>
#>     <div class="input-group">
#>       <label class="input-group-btn">
#>         <span class="btn btn-default btn-file">
#>           Browse...
#>           <input id="my_app-phewas" name="my_app-phewas" type="file" style="display: none;" accept="text/csv,text/comma-separated-values,text/plain,.csv"/>
#>         </span>
#>       </label>
#>       <input type="text" class="form-control" placeholder="No file selected" readonly="readonly"/>
#>     </div>
#>     <div id="my_app-phewas_progress" class="progress progress-striped active shiny-file-input-progress">
#>       <div class="progress-bar"></div>
#>     </div>
#>   </div>
#>   <div class="form-group shiny-input-container">
#>     <label class="control-label" for="my_app-genome">ID to SNP file</label>
#>     <div class="input-group">
#>       <label class="input-group-btn">
#>         <span class="btn btn-default btn-file">
#>           Browse...
#>           <input id="my_app-genome" name="my_app-genome" type="file" style="display: none;" accept="text/csv,text/comma-separated-values,text/plain,.csv"/>
#>         </span>
#>       </label>
#>       <input type="text" class="form-control" placeholder="No file selected" readonly="readonly"/>
#>     </div>
#>     <div id="my_app-genome_progress" class="progress progress-striped active shiny-file-input-progress">
#>       <div class="progress-bar"></div>
#>     </div>
#>   </div>
#>   <div class="form-group shiny-input-container">
#>     <label class="control-label" for="my_app-phenome">ID to phenome file</label>
#>     <div class="input-group">
#>       <label class="input-group-btn">
#>         <span class="btn btn-default btn-file">
#>           Browse...
#>           <input id="my_app-phenome" name="my_app-phenome" type="file" style="display: none;" accept="text/csv,text/comma-separated-values,text/plain,.csv"/>
#>         </span>
#>       </label>
#>       <input type="text" class="form-control" placeholder="No file selected" readonly="readonly"/>
#>     </div>
#>     <div id="my_app-phenome_progress" class="progress progress-striped active shiny-file-input-progress">
#>       <div class="progress-bar"></div>
#>     </div>
#>   </div>
#>   <h3>P-Value adjustment</h3>
#>   <p>Should multiple-comparisons adjustment be done on P-Values of phewas results file?</p>
#>   <div id="my_app-pval_correction" class="form-group shiny-input-radiogroup shiny-input-container">
#>     <label class="control-label" for="my_app-pval_correction">
#>       <h3>Correction type</h3>
#>     </label>
#>     <div class="shiny-options-group">
#>       <div class="radio">
#>         <label>
#>           <input type="radio" name="my_app-pval_correction" value="none" checked="checked"/>
#>           <span>None</span>
#>         </label>
#>       </div>
#>       <div class="radio">
#>         <label>
#>           <input type="radio" name="my_app-pval_correction" value="bonferroni"/>
#>           <span>Bonferroni</span>
#>         </label>
#>       </div>
#>       <div class="radio">
#>         <label>
#>           <input type="radio" name="my_app-pval_correction" value="BH"/>
#>           <span>Benjamini-Hochberg</span>
#>         </label>
#>       </div>
#>     </div>
#>   </div>
#> </div>
#> 
#>   </div>
#>   <div class = 'instructions-panel section'>
#>      
#> <h2>Preloaded data</h2>
#> 
#> <p>Supplied are snps that have been preloaded. Select a SNP in the dropdown and then click &#39;Use preloaded data&#39; to enter the app using the selected SNP.</p>
#> 
#> <h2>Input Data format</h2>
#> 
#> <p>There are three required files for loading data into the public facing version of MultimorbidityExplorer. All files are required to be in <code>.csv</code> format.</p>
#> 
#> <p>These files are: </p>
#> 
#> <p><strong>Phewas results file</strong>: This file contains the results of the (most likely) univariate statistical analysis correlating each phenotype code to your SNP or biomarker of interest. The columns are</p>
#> 
#> <ul>
#> <li><code>code</code>: character value uniquely denoting a given phenotype</li>
#> <li><code>OR</code>: Odds ratio from statistical test associating the given phenotype code with the biomarker of interest.</li>
#> <li><code>p_val</code>: The p-value associated with same test.</li>
#> <li><code>description</code>: Short description in words of what the code represents (E.g. <code>&quot;Heart Failure&quot;</code>).</li>
#> <li><code>category</code>: A hierarchical category denoting some grouping structure in your phenotypes. For instance all codes related to &#39;infectious diseases&#39;. These categories are used in coloring the manhattan plot.</li>
#> </ul>
#> 
#> <p>Any other columns that are added will be included in a small table available on mouseover of codes in app. </p>
#> 
#> <p><strong>ID to SNP file</strong>: Mapping between individual&#39;s IDs and the number of copies of the minor allele they have for the SNP of interest. A row is only needed if an individual has one or more copies of the minor allele (but data provided with rows for zero copies will work as well, just be less space efficient). </p>
#> 
#> <ul>
#> <li><code>IID</code>: Unique identifying character ID for each individual in your data. </li>
#> <li><code>snp</code>: Integer corresponding to the number of copies of the minor allele the individual possesses. </li>
#> </ul>
#> 
#> <p><strong>ID to phenome file</strong>: A mapping between an individuals ID and present phenotypes. If an individual has 10 present phenotypes they will have 10 rows in this csv; 25 phenotypes: 25 rows, etc.. Columns are: </p>
#> 
#> <ul>
#> <li><code>IID</code>: Unique identifying character ID for each individual (matches same column in <strong>ID to SNP file</strong>.)</li>
#> <li><code>code</code>: Unique identifying character ID for a given phenotype. This should match the column of the same name in <strong>Phewas results file</strong>.</li>
#> </ul>
#> 
#> 
#>   </div>
#> </div>
#>