Main Multimorbidity Explorer Dashboard: Server function

Generates a full dashboard page containing various visualizations for investigating comobidity patterns in individual level data and how they relate to the results of a phewas analysis.

main_dashboard(
  input,
  output,
  session,
  snp_name,
  phewas_results,
  individual_data,
  max_allowed_codes = 40,
  usage_instructions = "default",
  colors = list(light_grey = "#f7f7f7", med_grey = "#d9d9d9", light_blue = "#4292c6",
    green = "#74c476"),
  snp_colors,
  debug_mode = FALSE,
  show_back_button_messenger = NULL
)

Arguments

input, output, session
Auto-filled by callModule | ignore

snp_name
Character string containing the RSID of the snp you're viewing. Used to find annotation information.

phewas_results
Dataframe containing the results of the phewas study. Needs columns p_val, id, category(along with accompanying color), tooltip.

individual_data
Dataframe containing columns on IID, snp(# copies of allele), and columns for each code included.

max_allowed_codes
How many codes can the app show at any given time. Defaults to 40. (Too many and app may get slow.)

usage_instructions
HTML tags corresponding to static content to be displayed in bottom half of info panel. Any html content works. Defaults to light description.

colors
A list of CSS-valid colors to paint interface in if custom colors desired. Needs light_grey, med_grey, light_blue, light_blue, green.

snp_colors
Array of valid css color codes for 0, 1, and 2 copies of the minor allele in network plot.

debug_mode
Boolean controlling if changes in app state should be recorded in logs. Defaults to off.

Value

Shiny module of main Multimorbidity Explorer dashboard

See also

Examples

callModule(
  main_dashboard, 'main_dashboard',
  my_phewas_results,
  my_individual_data,
  usage_instructions = 'This app is complicated!'
)

#> Error in callModule(main_dashboard, "main_dashboard", my_phewas_results,     my_individual_data, usage_instructions = "This app is complicated!"): could not find function "callModule"