Subject-level network diagram: Server

Subject-level network diagram: Server

network_plot(
  input,
  output,
  session,
  network_data,
  highlighted_codes,
  snp_filter,
  viz_type = "free",
  update_freq = 15,
  action_object
)

Arguments

input, output, session
Auto-filled by callModule | ignore

network_data
Reactive object containing individual network data as generated by meToolkit::setup_network_data.

highlighted_codes
Reactive object containing list with type: ({'pattern', 'code'}) and codes: array of code names that comprise connection pattern to highlight in patients.

snp_filter
Reactive object containing boolean containing info on if we've filtered by snp or not.

viz_type
Character string containing info on which type of network we want to draw. "bipartite" for a plot that puts one node type on either size, or free for a traditional force directed layout. Defaults to 'free'.

update_freq
How many iterations of the layout simulation are run between redrawing the viz. Set to lower value for a smoother animation, higher for better performance. Default is 15 frames.

action_object
A reactiveVal that will be updated by the module upon isolation, deletion, or snp_filtering.

Value

Server component of interactive network plot. Returns type-payload list with the type "isolation, deletion, snp_filtering" to the passed action_object for updating app state.

See also

Examples

callModule(info_panel, 'info_panel', snp_name, individual_data, subset_maf)

#> Error in callModule(info_panel, "info_panel", snp_name, individual_data,     subset_maf): could not find function "callModule"