Combine and reconcile the three main data components needed for ME app and return in app-friendly format

Helper function used at the begining of a ME app that takes the raw format of the three data sets required, phewas results, individual snp copies, and individual phenotype pairs, and combines them into a list object that combines individual data and checks phewas results along with extracting the snp name from the column of the genotype file.

reconcile_data(phewas_results, id_to_snp, id_to_code)

Arguments

phewas_results
Dataframe containing the results of the phewas study. Needs columns p_val, id, category(along with accompanying color), tooltip.

id_to_snp
Dataframe containing columns on IID, snp(# copies of allele).

id_to_code
Dataframe containing column IID, and columns for each code included.

Value

List with snp_name, individual_data, and phewas_results properties for use in ME app.

Examples

reconcile_data(phewas_table, id_to_snp, phenotype_id_pairs)

#> Error in is.data.frame(x): object 'phewas_table' not found