Generate sample individual level data based upon some phewas results

This function will take the results of a phewas study (most importantly a code, case and control numbers, and an OR) and simulate as many individual datapoints as you desire.

simIndividualData(phewas_results, num_patients, snp_prev)

Arguments

phewas_results
Phewas results. Needs a column code with the id of a given phecode, code_proportion: with the proportion of individuals that have the code in the entire population, and OR: the odds ratio of having a code given having the snp of interest.

num_patients
How many patients should be simulated.

snp_prev
The overall population prevalence of the snp you want to simulate.

Value

list with following components

snp_status

tibble with patient ID and snp status (snp)

phenotypes

wide tibble with an ID column and a column for each code provided in the phewas_resuts input. Cell values are if the patient had that given code.

Examples

simIndividualData(fake_phewas_results, 10, 0.15)

#> $snp_status
#> # A tibble: 10 x 2
#>    id      snp
#>    <chr> <int>
#>  1 r1        0
#>  2 r10       0
#>  3 r2        0
#>  4 r3        0
#>  5 r4        0
#>  6 r5        0
#>  7 r6        0
#>  8 r7        0
#>  9 r8        0
#> 10 r9        0
#> 
#> $phenotypes
#> # A tibble: 10 x 1,579
#>    id    `008.00` `008.50` `008.52` `008.60` `010.00` `031.00` `038.00` `038.10`
#>    <chr>    <int>    <int>    <int>    <int>    <int>    <int>    <int>    <int>
#>  1 r1           0        0        0        0        0        0        0        0
#>  2 r10          0        0        0        0        0        0        0        0
#>  3 r2           0        1        0        0        0        0        0        0
#>  4 r3           0        0        0        0        0        0        0        0
#>  5 r4           0        0        0        0        0        0        0        0
#>  6 r5           0        0        0        0        0        0        0        0
#>  7 r6           0        0        0        0        0        0        0        0
#>  8 r7           0        0        0        0        0        0        0        0
#>  9 r8           0        0        0        0        0        0        0        0
#> 10 r9           0        0        0        0        0        0        0        0
#> # … with 1,570 more variables: `038.20` <int>, `038.30` <int>, `041.00` <int>,
#> #   `041.10` <int>, `041.11` <int>, `041.12` <int>, `041.20` <int>,
#> #   `041.40` <int>, `041.80` <int>, `041.90` <int>, `053.00` <int>,
#> #   `053.10` <int>, `054.00` <int>, `070.00` <int>, `070.20` <int>,
#> #   `070.30` <int>, `070.40` <int>, `070.90` <int>, `071.00` <int>,
#> #   `071.10` <int>, `078.00` <int>, `079.00` <int>, `079.20` <int>,
#> #   `080.00` <int>, `081.00` <int>, `081.10` <int>, `081.11` <int>,
#> #   `081.12` <int>, `090.00` <int>, `110.00` <int>, `110.10` <int>,
#> #   `110.11` <int>, `110.12` <int>, `110.13` <int>, `110.20` <int>,
#> #   `112.00` <int>, `112.30` <int>, `117.00` <int>, `117.10` <int>,
#> #   `117.40` <int>, `130.00` <int>, `132.00` <int>, `136.00` <int>,
#> #   `145.00` <int>, `145.20` <int>, `145.30` <int>, `145.40` <int>,
#> #   `145.50` <int>, `149.00` <int>, `149.10` <int>, `149.20` <int>,
#> #   `149.30` <int>, `149.40` <int>, `149.50` <int>, `149.90` <int>,
#> #   `150.00` <int>, `151.00` <int>, `153.00` <int>, `153.20` <int>,
#> #   `153.30` <int>, `155.00` <int>, `155.10` <int>, `157.00` <int>,
#> #   `158.00` <int>, `159.00` <int>, `159.20` <int>, `159.30` <int>,
#> #   `159.40` <int>, `164.00` <int>, `165.00` <int>, `165.10` <int>,
#> #   `170.00` <int>, `170.10` <int>, `170.20` <int>, `172.00` <int>,
#> #   `172.10` <int>, `172.11` <int>, `172.20` <int>, `172.21` <int>,
#> #   `172.22` <int>, `172.30` <int>, `173.00` <int>, `174.00` <int>,
#> #   `174.10` <int>, `174.11` <int>, `174.30` <int>, `175.00` <int>,
#> #   `180.00` <int>, `180.10` <int>, `180.30` <int>, `182.00` <int>,
#> #   `184.00` <int>, `184.10` <int>, `184.11` <int>, `184.20` <int>,
#> #   `185.00` <int>, `187.00` <int>, `187.10` <int>, `187.20` <int>,
#> #   `189.00` <int>, …
#>