Generate sample individual level data based upon a PheWAS results set

This function will take the results of a PheWAS study (most importantly a code, case and control numbers, and an OR) and simulate as many individual datapoints as you desire. Should only be used for testing of apps and not any real simulation studies.

sim_individual_data(phewas_results, num_patients, snp_prev)

Arguments

phewas_results
Phewas results. Needs a column code with the id of a given phecode, and OR: the odds ratio of having a code given having the snp of interest. Optionally, individuals that have the code in the entire population can be provided with column code_proportion.

num_patients
How many patients should be simulated.

snp_prev
The overall population prevalence of the snp you want to simulate.

Value

list with following components

snp_status

tibble with patient ID and snp status (snp)

phenotypes

wide tibble with an ID column and a column for each code provided in the phewas_resuts input. Cell values are if the patient had that given code.

Examples

simulated_phewas <- sim_phewas_results(n_codes = 100, n_categories = 11)
sim_individual_data(simulated_phewas, 10, 0.15)

#> Warning: NAs produced

#> $snp_status
#> # A tibble: 10 x 2
#>    id      snp
#>    <chr> <int>
#>  1 r1        0
#>  2 r10       0
#>  3 r2        1
#>  4 r3        0
#>  5 r4        1
#>  6 r5        1
#>  7 r6        0
#>  8 r7        0
#>  9 r8        0
#> 10 r9        0
#> 
#> $phenotypes
#> # A tibble: 10 x 101
#>    id    `001.00` `002.00` `003.00` `004.00` `005.00` `006.00` `007.00` `008.00`
#>    <chr>    <int>    <int>    <int>    <int>    <int>    <int>    <int>    <int>
#>  1 r1           1        0        1        0        0        0        0        1
#>  2 r10          0        0        1        0        0        0        0        0
#>  3 r2          NA        1       NA       NA        0        1        1        1
#>  4 r3           0        0        0        0        0        0        1        0
#>  5 r4          NA        0       NA       NA        1        1        0        0
#>  6 r5          NA        1       NA       NA        1        0        1        1
#>  7 r6           0        0        1        0        0        0        0        0
#>  8 r7           0        0        0        0        0        1        0        1
#>  9 r8           0        0        0        0        0        0        0        0
#> 10 r9           1        1        0        0        0        0        0        0
#> # … with 92 more variables: `009.00` <int>, `010.00` <int>, `011.00` <int>,
#> #   `012.00` <int>, `013.00` <int>, `014.00` <int>, `015.00` <int>,
#> #   `016.00` <int>, `017.00` <int>, `018.00` <int>, `019.00` <int>,
#> #   `020.00` <int>, `021.00` <int>, `022.00` <int>, `023.00` <int>,
#> #   `024.00` <int>, `025.00` <int>, `026.00` <int>, `027.00` <int>,
#> #   `028.00` <int>, `029.00` <int>, `030.00` <int>, `031.00` <int>,
#> #   `032.00` <int>, `033.00` <int>, `034.00` <int>, `035.00` <int>,
#> #   `036.00` <int>, `037.00` <int>, `038.00` <int>, `039.00` <int>,
#> #   `040.00` <int>, `041.00` <int>, `042.00` <int>, `043.00` <int>,
#> #   `044.00` <int>, `045.00` <int>, `046.00` <int>, `047.00` <int>,
#> #   `048.00` <int>, `049.00` <int>, `050.00` <int>, `051.00` <int>,
#> #   `052.00` <int>, `053.00` <int>, `054.00` <int>, `055.00` <int>,
#> #   `056.00` <int>, `057.00` <int>, `058.00` <int>, `059.00` <int>,
#> #   `060.00` <int>, `061.00` <int>, `062.00` <int>, `063.00` <int>,
#> #   `064.00` <int>, `065.00` <int>, `066.00` <int>, `067.00` <int>,
#> #   `068.00` <int>, `069.00` <int>, `070.00` <int>, `071.00` <int>,
#> #   `072.00` <int>, `073.00` <int>, `074.00` <int>, `075.00` <int>,
#> #   `076.00` <int>, `077.00` <int>, `078.00` <int>, `079.00` <int>,
#> #   `080.00` <int>, `081.00` <int>, `082.00` <int>, `083.00` <int>,
#> #   `084.00` <int>, `085.00` <int>, `086.00` <int>, `087.00` <int>,
#> #   `088.00` <int>, `089.00` <int>, `090.00` <int>, `091.00` <int>,
#> #   `092.00` <int>, `093.00` <int>, `094.00` <int>, `095.00` <int>,
#> #   `096.00` <int>, `097.00` <int>, `098.00` <int>, `099.00` <int>,
#> #   `100.00` <int>
#>