Generate sample individual level data based upon some phewas results

This function will take the results of a phewas study (most importantly a code, case and control numbers, and an OR) and simulate as many individual datapoints as you desire.

simIndividualData(phewas_results, num_patients, snp_prev)

Arguments

phewas_results
Phewas results. Needs a column code with the id of a given phecode, code_proportion: with the proportion of individuals that have the code in the entire population, and OR: the odds ratio of having a code given having the snp of interest.

num_patients
How many patients should be simulated.

snp_prev
The overall population prevalence of the snp you want to simulate.

Value

list with following components

snp_status

tibble with patient ID and snp status (snp)

phenotypes

wide tibble with an ID column and a column for each code provided in the phewas_resuts input. Cell values are if the patient had that given code.

Examples

simIndividualData(fake_phewas_results, 10, 0.15)

#> Error in n(): could not find function "n"